requirements.txt
fastapi==0.78.0
uvicorn==0.15.0
python-telegram-bot==20.0
requests==2.25.1
apscheduler==3.10.0
# openai==0.27.7 